src/requirements.txt
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
ansible==11.0.0
ansible-core==2.18.0
anyio @ file:///croot/anyio_1729121277521/work
astor @ file:///work/perseverance-python-buildout/croot/astor_1698866721323/work
beautifulsoup4==4.12.3
blinker==1.9.0
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
fire==0.7.0
Flask==3.1.0
fonttools==4.55.1
h11 @ file:///croot/h11_1706652277403/work
httpcore @ file:///croot/httpcore_1706728464539/work
httpx @ file:///croot/httpx_1723474802858/work
idna==3.10
imageio==2.36.1
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
networkx @ file:///croot/networkx_1720002482208/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp312-cp312-linux_x86_64.whl#sha256=1d700f51d8b4fa684d858c9e3b56b1656bc5c82b6b79ff08d4e3b491c430059f
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging==24.2
paddleocr==2.9.1
paddlepaddle==3.0.0b2
pillow @ file:///croot/pillow_1731594689713/work
protobuf==4.25.3
pyclipper==1.3.0.post6
pydantic==2.10.2
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
PyYAML==6.0.2
RapidFuzz==3.10.1
requests==2.32.3
resolvelib==1.0.1
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
shapely==2.0.6
six==1.16.0
sniffio @ file:///croot/sniffio_1705431295498/work
soupsieve==2.6
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.44.0
